% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_log.R
\name{analyze_log}
\alias{analyze_log}
\title{Generate a table of the log file for all quizzes in the course}
\usage{
analyze_log(courselocation)
}
\arguments{
\item{courselocation}{parent folder for the course}
}
\value{
If things went well, a dataframe containing a summary of all quizzes.
Otherwise an error message is returned.
}
\description{
This function creates a dataframe summarizing the quizzes and submissions in the course.
This is used by the quizmanager part of the package
}
