% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_overview.R
\name{analyze_overview}
\alias{analyze_overview}
\title{Generate a summary of all quizzes in the course}
\usage{
analyze_overview(courselocation)
}
\arguments{
\item{courselocation}{parent folder for the course}
}
\value{
If things went well, a dataframe containing a summary of all quizzes.
}
\description{
This function creates a dataframe summarizing the quizzes and submissions in the course.
This is a helper function called by the quizmanager app.
}
