% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_quiz.R
\name{analyze_quiz}
\alias{analyze_quiz}
\title{Generate a table showing detailed submission results for a selected quiz}
\usage{
analyze_quiz(courselocation, selected_quiz)
}
\arguments{
\item{courselocation}{parent folder for the course}

\item{selected_quiz}{quiz to be analyzed}
}
\value{
If things went well, a dataframe containing detailed information
for each question in a quiz.
Otherwise an error message is returned.
}
\description{
This function creates a dataframe showing
what each student submitted for a quiz, and some summary statistics
for each question.
This is used by the quizmanager part of the package
}
