% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_quizzes.R
\name{analyze_quizzes}
\alias{analyze_quizzes}
\title{Generate a summary of all quizzes in the course}
\usage{
analyze_quizzes(
  courselocation,
  grades_type = "overview",
  duedate_filter = FALSE
)
}
\arguments{
\item{courselocation}{parent folder for the course}

\item{grades_type}{one of c("overview", "student", "question")}

\item{duedate_filter}{logical indicating whether to output statistics for past quizzes only}
}
\value{
If things went well, a dataframe containing a summary of all quizzes.
Otherwise an error message is returned.
}
\description{
This function creates a dataframe summarizing the quizzes in the course.
This is used by the quizmanager part of the package
}
