% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_scoretable.R
\name{analyze_scoretable}
\alias{analyze_scoretable}
\title{Generate a score table for all quizzes in the course}
\usage{
analyze_scoretable(courselocation)
}
\arguments{
\item{courselocation}{parent folder for the course}
}
\value{
If things went well, a dataframe containing a table with scores for all students and quizzes.
}
\description{
This function creates a dataframe showing scores for each quiz and student in the course.
For multiple allowed attempts, only the highest score is shown.
This table is useful for grading, less so for analysis of individual students or quizzes.
This is a helper function called by the quizmanager app.
}
