% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_student.R
\name{analyze_student}
\alias{analyze_student}
\title{Generate a table showing detailed submission results for a selected student}
\usage{
analyze_student(courselocation, selected_student)
}
\arguments{
\item{courselocation}{parent folder for the course}

\item{selected_student}{student to be analyzed}
}
\value{
If things went well, a dataframe containing detailed information
for the selected student.
Otherwise an error message is returned.
}
\description{
This function creates a dataframe showing
a detailed submission/performance history for a selected student.
This is used by the quizmanager part of the package
}
