% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_courselocation.R
\name{check_courselocation}
\alias{check_courselocation}
\title{Check that the provided course location is a valid main quizgrader folder}
\usage{
check_courselocation(courselocation)
}
\arguments{
\item{courselocation}{location of course}
}
\value{
If everything is right, NULL is returned
If anything is not right, an error message is returned
}
\description{
This function checks that a valid quiagrader folder is provided as path
}
