% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission.R
\name{check_file}
\alias{check_file}
\title{Check submission for errors before grading}
\usage{
check_file(submission, solution, studentid, quizid, classlist)
}
\arguments{
\item{submission}{The data frame containing the student submission}

\item{solution}{The data frame containing the solution}

\item{studentid}{The student who's submission is checked}

\item{quizid}{The ID of the quiz to be checked}

\item{courselist}{A data frame containing due dates, previous submissions, etc.
is to be created}
}
\value{
If anything is not right, an error message is returned.
Otherwise, NULL is returned, meaning all checks went ok.
}
\description{
This function checks the submitted quiz
performs some checks to make sure the formatting is right
and also checks to ensure due date or number of submissions has not been exceeded
}
