% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{Check that information submitted by students through UI is correct}
\usage{
check_metadata(metadata, studentlist)
}
\arguments{
\item{metadata}{list containing elements StudentID and Password}

\item{gradelist}{Data frame containing the grade list}
}
\value{
An error message if user input is not valid
otherwise NULL
}
\description{
This function takes the metadata submitted by a student
(Name, ID, Password) and checks it against information
in the grade tracking list to ensure that student is present
and the submitted information valid
}
