% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quiz.R
\name{check_quiz}
\alias{check_quiz}
\title{Check that a data frame sent to this function has a valid structure for a complete quiz}
\usage{
check_quiz(quizdf)
}
\arguments{
\item{quizdf}{Data frame of quiz that needs to be checked}
}
\value{
An error message if it's not a valid quiz data frame,
otherwise NULL
}
\description{
This function takes the data frame or tibble
that is supposed to contain a valid completed quiz
and checks that structure and content are valid
}
