% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_studentlist.R
\name{check_studentlist}
\alias{check_studentlist}
\title{Check the studentlist file}
\usage{
check_studentlist(studentdf)
}
\arguments{
\item{studentdf}{dataframe with student list information}
}
\value{
If everything is right, NULL is returned
If anything is not right, an error message is returned
}
\description{
This function checks a studentlist dataframe to ensure it is valid
}
