% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submission.R
\name{check_submission}
\alias{check_submission}
\title{Check that a data frame sent to this function has a valid structure for student submitted quiz}
\usage{
check_submission(quizdf)
}
\arguments{
\item{quizdf}{Data frame of quiz that needs to be checked}
}
\value{
An error message if it's not a valid quiz data frame,
otherwise NULL
}
\description{
This function takes the data frame or tibble
that is supposed to contain a valid quiz submission
and checks that structure and content are valid
}
