% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_submission_logs.R
\name{compile_submission_logs}
\alias{compile_submission_logs}
\title{Compile log files from each quiz submission}
\usage{
compile_submission_logs(
  studentid,
  studentsubmissions_folder,
  df_format = "condensed"
)
}
\arguments{
\item{studentid}{id of student(s)}

\item{studentsubmissions_folder}{folder where student quiz submissions are saved}

\item{df_format}{which data frame format to output: "condensed" for a single row per student per quiz; "byquestion" for question-specific rows}
}
\value{
If things went well, a dataframe containing student quiz results from all submitted quizzes.
Otherwise an error message is returned.
}
\description{
This function creates a dataframe from the log files generated during student submission (adaptation of the displayed results table).
}
