% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_submissions.R
\name{compile_submissions}
\alias{compile_submissions}
\title{Generate a summary of all quizzes in the course}
\usage{
compile_submissions(courselocation)
}
\arguments{
\item{courselocation}{parent folder for all course directories}
}
\value{
If things went well, a dataframe containing a summary of all quizzes.
Otherwise an error message is returned.
}
\description{
This function creates a dataframe summarizing the quizzes in the course.
}
