% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_student_stats.R
\name{compute_student_stats}
\alias{compute_student_stats}
\title{Compute overall quiz stats for each student}
\usage{
compute_student_stats(studentid, quizid, gradelist)
}
\arguments{
\item{studentid}{id of student}

\item{quizid}{id of quiz}

\item{gradelist}{gradelist data frame}
}
\value{
If things went well, a vector containing student stats.
Otherwise an error message is returned.
}
\description{
This function
}
