% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gradelist.R
\name{create_gradelist}
\alias{create_gradelist}
\title{Create main quiz tracking file}
\usage{
create_gradelist(courselocation)
}
\arguments{
\item{courselocation}{Path to course}
}
\value{
A data frame with each student as row, and columns
containing student information and information for each quiz.
}
\description{
This function takes the user-supplied
list of students and a directory containing all quizzes
and generates a data frame that has for each student
their information needed for login and multiple columns
for each quiz tracking/recording student submission
}
