% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_student_quizzes.R
\name{create_student_quizzes}
\alias{create_student_quizzes}
\title{Create student quizzes from full quiz sheets}
\usage{
create_student_quizzes(courselocation)
}
\arguments{
\item{courselocation}{Path to main course directory}
}
\value{
errormsg Null if quiz generation was successful,
otherwise contains an error message
}
\description{
This function takes all complete quiz Excel sheets.
It checks each quiz sheet for proper formatting.
It then strips some of the content and produces quiz sheets
to be given to students for filling out.
It places those quiz sheets into the student_quiz_sheets folder
}
