% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_quiz.R
\name{grade_quiz}
\alias{grade_quiz}
\title{Grade a quiz}
\usage{
grade_quiz(submission, solution)
}
\arguments{
\item{submission}{The data frame containing the student submission}

\item{solution}{The data frame containing the solution
is to be created}
}
\value{
If things went well, a data frame containing "Correct" or "Not correct" for each question.
Otherwise an error message is returned.
}
\description{
This function takes a data frame containing student submissions
and a data frame containing the answers and evaluates/grades.
The specific grading depends on the type of question.
}
