% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_logfile.R
\name{load_logfile}
\alias{load_logfile}
\title{Find and load the most recent log file}
\usage{
load_logfile(courselocation)
}
\arguments{
\item{courselocation}{parent folder for the course}
}
\value{
If things went well, a dataframe containing the entries of the latest log file.
Otherwise an error message is returned.
}
\description{
Determines the most recent log file at time of function call
and loads it.
This is used by both the quizgrader and quizmanager parts of the package
}
