% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_course_list.R
\name{make_course_list}
\alias{make_course_list}
\title{Create main quiz tracking data frame}
\usage{
make_course_list(courselistfile, quizlocation)
}
\arguments{
\item{courselistfile}{Name of file that contains the course list.
The file needs to be an Excel file with at least
columns Lastname, Firstname, ID and Password.
It should be based on the provided template.}

\item{quizlocation}{The name of the folder where
the complete quizzes are stored.
Those quizzes will be parsed to generate the data frame.}
}
\value{
A data frame with each student as row, and columns
containing student information and information for each quiz.
}
\description{
This function takes the user-supplied
list of students and a directory containing all quizzes
and generates a data frame that has for each student
their information needed for login and multiple columns
for each quiz tracking/recording student submission
}
