% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizgrader-package.R
\docType{package}
\name{quizgrader-package}
\alias{quizgrader}
\alias{quizgrader-package}
\title{quizgrader: A package to do grading}
\description{
This package allows easy setup and deployment of autograded online quizzes.
}
\details{
Start the main menu of the package by calling quizmanager().

To learn more about how to use the package, see the documentation
on the package website:
https://andreashandel.github.io/quizgrader/
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andreashandel.github.io/quizgrader/}
  \item \url{https://github.com/andreashandel/quizgrader/}
  \item Report bugs at \url{https://github.com/ahgroup/DSAIDE/issues}
}

}
\author{
\strong{Maintainer}: Andreas Handel \email{ahandel@uga.edu} (\href{https://orcid.org/0000-0002-4622-1146}{ORCID})

Other contributors:
\itemize{
  \item Cody Dailey [contributor]
  \item Brian McKay [contributor]
}

}
\keyword{internal}
