% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizgrader.R
\docType{package}
\name{quizgrader}
\alias{quizgrader}
\title{quizgrader: A package to do grading}
\description{
More information text here
}
\details{
Start the main menu of the package by calling quizmanager().

To learn more about how to use the package, see the documentation
on the package website:
https://andreashandel.github.io/quizgrader/
}
