% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_courselist.R
\name{read_classlist}
\alias{read_classlist}
\title{Load and read the courselist file}
\usage{
read_classlist(courselist_folder)
}
\arguments{
\item{courselist_folder}{Name of folder where all the courselist files are stored}
}
\value{
Returns the courselist as data frame
If anything is not right, an error message is returned.
}
\description{
This function loads the courselist and does some formatting
}
