% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gradelist.R
\name{read_gradelist}
\alias{read_gradelist}
\title{Load and read the gradelist file}
\usage{
read_gradelist(gradelist_folder)
}
\arguments{
\item{gradelist_folder}{Name of folder where all the courselist files are stored}
}
\value{
Returns the gradelist as data frame
If anything is not right, an error message is returned.
}
\description{
This function loads the gradelist file and does some formatting
}
