% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_studentlist.R
\name{read_studentlist}
\alias{read_studentlist}
\title{Load and read the studentlist file}
\usage{
read_studentlist(studentlist_folder)
}
\arguments{
\item{studentlist_folder}{Name of folder where the studentlist file is stored}
}
\value{
Returns the studentlist as data frame
If anything is not right, an error message is returned.
}
\description{
This function loads the studentlist file and does some formatting
}
