% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_submissions_log.R
\name{read_submissions_log}
\alias{read_submissions_log}
\title{Load and read the submissions log file}
\usage{
read_submissions_log(studentsubmissions_folder)
}
\arguments{
\item{studentsubmissions_folder}{Name of folder where all the student submission files are stored}
}
\value{
Returns the submissions log file as data frame
If anything is not right, an error message is returned.
}
\description{
This function loads the submissions log file and does some formatting
}
