% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_grade.R
\name{save_grade}
\alias{save_grade}
\title{Save the student grade to the gradelist file}
\usage{
save_grade(score, studentid, quizid, gradelist, gradelists_folder)
}
\arguments{
\item{score}{student score}

\item{studentid}{student id}

\item{quizid}{quiz id}

\item{gradelist}{gradelist data frame}

\item{gradelists_folder}{folder to save gradelists}
}
\value{
An error message if it's not a valid quiz data frame,
otherwise NULL
}
\description{
This function takes the data frame or tibble
that is supposed to contain a valid completed quiz
and checks that structure and content are valid
}
