% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_course_shiny.R
\name{start_course_shiny}
\alias{start_course_shiny}
\title{Set up structure for a new course through shiny}
\usage{
start_course_shiny(coursename, courselocation = NULL)
}
\value{
The script creates folders.
If things went well, a message is returned stating where the new course was created.
Otherwise an error message is returned.
}
\description{
This function is called from the main shiny menu
it asks for a course name and a location for the course
it then sends it on to the start_course function
which creates template folders and files for quizzes
to be filled by the teacher
}
