% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_course.R
\name{summarize_course}
\alias{summarize_course}
\title{Generate a summary of all quizzes in the course}
\usage{
summarize_course(courselocation)
}
\arguments{
\item{courselocation}{parent folder for all course directories}
}
\value{
If things went well, a dataframe containing a summary of all quizzes.
Otherwise an error message is returned.
}
\description{
This function creates a dataframe summarizing the quizzes in the course.
}
