% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quiz.R
\name{check_quiz}
\alias{check_quiz}
\title{Check that a file sent to this function has a valid structure for a complete quiz}
\usage{
check_quiz(quizfile)
}
\arguments{
\item{quizfile}{Path and name of file to be checked}
}
\value{
An error message if it's not a valid quiz,
otherwise NULL
}
\description{
This function takes the name/location of a
file and tries to load it and check that it has a valid
structure and content for a quiz
}
