% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{Create file and folder zip file for deployment}
\usage{
make_package(courselocation, newpackage = TRUE)
}
\arguments{
\item{courselocation}{Path to main course directory}

\item{newpackage}{If true, all files needed for initial setup are copied, otherwise only those needed for updates}
}
\value{
errormsg Null if quiz generation was successful,
otherwise contains an error message
}
\description{
This function takes all files and folders
needed for the server and places it in a zip file.
The user then needs to copy the files to the server
and unzip them in the app directory.
}
