% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizgrader.R
\docType{package}
\name{quizgrader-package}
\alias{quizgrader}
\alias{quizgrader-package}
\title{quizgrader: Automated managament and grading of quizzes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Allows generation of quizzes using simple spreadsheets, student submit online and quizzes are autograded.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andreashandel.github.io/quizgrader/}
  \item \url{https://github.com/andreashandel/quizgrader/}
  \item Report bugs at \url{https://github.com/ahgroup/DSAIDE/issues}
}

}
\author{
\strong{Maintainer}: Andreas Handel \email{ahandel@uga.edu} (\href{https://orcid.org/0000-0002-4622-1146}{ORCID})

Other contributors:
\itemize{
  \item Cody Dailey [contributor]
  \item Brian McKay [contributor]
}

}
\keyword{internal}
