% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizmanager.R
\name{quizmanager}
\alias{quizmanager}
\title{The main menu for creating and administering quizzes}
\usage{
quizmanager(courselocation = NULL)
}
\arguments{
\item{courselocation}{Path to main course directory}
}
\description{
This function opens a Shiny app called quizmanager_app with a menu
that allows the user to perform all relevant quiz management operations.
}
\details{
Run this function with no arguments to start the main menu (a Shiny app).
You can also provide a path to an existing course directory.
}
\examples{
\dontrun{quizmanager()}
}
\author{
Andreas Handel
}
