% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_studentlist.R
\name{read_studentlist}
\alias{read_studentlist}
\title{Load and check the studentlist file}
\usage{
read_studentlist(courselocation)
}
\arguments{
\item{courselocation}{Path to course}
}
\value{
the studentlist as data frame
If anything is not right, an error message is returned.
}
\description{
This function loads the user generated studentlist and checks it
}
