% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_grade.R
\name{save_grade}
\alias{save_grade}
\title{Save the student grade to the gradelist file}
\usage{
save_grade(score, gradelist_folder)
}
\arguments{
\item{score}{student score}

\item{gradelist_folder}{folder where gradelist is stored}
}
\value{
An error message if it's not a valid quiz data frame,
otherwise NULL
}
\description{
This function takes the data frame or tibble
that is supposed to contain a valid completed quiz
and checks that structure and content are valid
}
