% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_course.R
\name{start_course}
\alias{start_course}
\title{Set up structure for a new course}
\usage{
start_course(coursename, courselocation = NULL)
}
\arguments{
\item{coursename}{Name of course to be created
this will be the main folder name
and will contain sub-folders for all the materials}

\item{courselocation}{The location where the new course folder
is to be created}
}
\value{
The function creates folders an files for the new course.
The function returns a list containing error or success information.
If things went well, the list element status is 0 and a the list
element message is a string specifying the location of the new course folder.
Otherwise an error status of 1 and an error message is returned.
}
\description{
This function creates a new course/project
and creates template folders and files for quizzes
to be filled by the teacher
}
